tempo/requirements.txt
playwright
torch
torchvision
transformers
pillow
